Move script classes into new namespace
