+++ lib/Elastica/Bulk.php
@@ -7,6 +7,7 @@ use Elastica\Bulk\Response as BulkResponse;
 use Elastica\Bulk\ResponseSet;
 use Elastica\Exception\Bulk\ResponseException as BulkResponseException;
 use Elastica\Exception\InvalidException;
+use Elastica\Script\AbstractScript;
 
 class Bulk
 {
@@ -190,8 +191,8 @@ class Bulk
     }
 
     /**
-     * @param \Elastica\AbstractScript $script
-     * @param string                   $opType
+     * @param \Elastica\Script\AbstractScript $script
+     * @param string                          $opType
      *
      * @return $this
      */
@@ -218,8 +219,8 @@ class Bulk
     }
 
     /**
-     * @param \Elastica\AbstractScript|\Elastica\Document|array $data
-     * @param string                                            $opType
+     * @param \Elastica\Script\AbstractScript|\Elastica\Document|array $data
+     * @param string                                                   $opType
      *
      * @return $this
      */
