+++ test/lib/Elastica/Test/Aggregation/MaxTest.php
@@ -4,7 +4,7 @@ namespace Elastica\Test\Aggregation;
 use Elastica\Aggregation\Max;
 use Elastica\Document;
 use Elastica\Query;
-use Elastica\Script;
+use Elastica\Script\Script;
 
 class MaxTest extends BaseAggregationTest
 {
