+++ test/lib/Elastica/Test/Aggregation/ScriptTest.php
@@ -4,7 +4,7 @@ namespace Elastica\Test\Aggregation;
 use Elastica\Aggregation\Sum;
 use Elastica\Document;
 use Elastica\Query;
-use Elastica\Script;
+use Elastica\Script\Script;
 
 class ScriptTest extends BaseAggregationTest
 {
