+++ test/lib/Elastica/Test/QueryTest.php
@@ -7,8 +7,8 @@ use Elastica\Query;
 use Elastica\Query\Builder;
 use Elastica\Query\Term;
 use Elastica\Query\Text;
-use Elastica\Script;
-use Elastica\ScriptFields;
+use Elastica\Script\Script;
+use Elastica\Script\ScriptFields;
 use Elastica\Suggest;
 use Elastica\Test\Base as BaseTest;
 use Elastica\Type;
